Refactor TemporalVectorIndex into modular structure
